+++ src/components/CreateWO/createWO.scss
@@ -1,27 +1,9 @@
-@import "../index.scss";
+@import "../../index.scss";
 
 .home {
 	.showViews {
 		.views {
 			height: calc(100vh - 93px);
-		}
-		.view1 {
-			justify-content: center;
-			align-items: center;
-			display: flex;
-			flex-direction: column;
-			.nombreInge {
-				opacity: 1;
-				transition: opacity 1.5s ease-in-out;
-			}
-			.btnValidate {
-				margin-top: 1em;
-			}
-			.version {
-				font-size: 0.7em;
-				position: absolute;
-				bottom: 2em;
-			}
 		}
 		.view2 {
 			justify-content: center;
