+++ src/components/CreateWO/views/View6Refacciones.jsx
