+++ src/components/Layout/layout.scss
@@ -6,7 +6,7 @@
 	.nav-bar {
 		width: 100vw;
 		display: flex;
-		background-color: #000000;
+		background-color: #202020;
 		.logo {
 			background-image: url("../../assets/images/icons/icon-384x384.png");
 			background-repeat: no-repeat;
