+++ src/components/Layout/sso.scss
@@ -1,0 +1,56 @@
+.ssoC {
+	height: calc(100vh - 93px);
+	.logged {
+		display: grid;
+		grid-template-rows: 1fr 2fr;
+		row-gap: 1em;
+		height: 100%;
+		.ingeData {
+			background-image: url("../../assets/images/userBG.gif");
+			background-size: cover;
+			p {
+				color: white;
+				font-weight: bold;
+				font-size: larger;
+				margin-left: 1em;
+			}
+			.logoutIcon {
+				color: white;
+				position: absolute;
+				right: 1rem;
+				top: 50px;
+				font-size: 2em;
+			}
+		}
+		.tasks {
+			display: grid;
+			grid-template-rows: 1fr 1fr;
+			.taskBtn {
+				margin: auto;
+				height: 4em;
+				width: 16em;
+				font-size: 1.1em;
+				i {
+					margin-right: 0.4em;
+					font-size: 2em;
+				}
+			}
+		}
+	}
+	.notLogged {
+		justify-content: center;
+		align-items: center;
+		display: flex;
+		flex-direction: column;
+		.btnValidate {
+			margin-top: 1em;
+		}
+	}
+	.version {
+		font-size: 0.8em;
+		// position: absolute;
+		// bottom: 2em;
+		display: flex;
+		justify-content: center;
+	}
+}
