+++ src/index.scss
@@ -1,46 +1,47 @@
 @font-face {
-  font-family: inspira;
-  // src: url("Inspira.ttf"); //para codesandbox
-  src: url("./assets/Inspira.ttf"); //para vscode
+	font-family: inspira;
+	// src: url("Inspira.ttf"); //para codesandbox
+	src: url("./assets/Inspira.ttf"); //para vscode
 }
 
 body {
-  // overflow: hidden;
-  padding: 0;
-  margin: 0;
-  background-color: #f5f5f5;
-  font-family: inspira, -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto",
-    "Oxygen", "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
-    sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
+	// overflow: hidden;
+	padding: 0;
+	margin: 0;
+	background-color: #f5f5f5;
+	font-family: inspira, -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto",
+		"Oxygen", "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
+		sans-serif;
+	-webkit-font-smoothing: antialiased;
+	-moz-osx-font-smoothing: grayscale;
 }
 * {
-  box-sizing: border-box;
+	box-sizing: border-box;
 }
 a {
-  padding: 10px 15px;
-  margin: 0 5px;
-  cursor: pointer;
-  text-decoration: none;
+	padding: 10px 15px;
+	margin: 0 5px;
+	cursor: pointer;
+	text-decoration: none;
 }
 /* Colors for theme */
 // $primary: #57acf1; //gehrsxtra
-$primary: #595959; //semantic linkedin button
+// $primary: #595959; //semantic linkedin button
+$primary: #444444; //semantic linkedin button
 // $secondary: #00d4ff;
 // $secondary: #54c8ff;
 $secondary: #295a84;
-$grayed: #295a84;
+$grised: #202020;
 // $grayed: #c5d0da;
 
 :export {
-  primary: $primary;
-  secondary: $secondary;
-  grayed: $grayed;
+	primary: $primary;
+	secondary: $secondary;
+	grised: $grised;
 }
 
 @media print {
-  body {
-    background-color: white !important;
-  }
+	body {
+		background-color: white !important;
+	}
 }
